facebook-reg-date-calc/solution-py3/requirements.txt
python-dateutil
git+https://github.com/mobolic/facebook-sdk.git
pytest

